Removed babel dependency.
This commit officially work without the babel dependency. However
after doing some digging and researching, I came to the realization
that babel may just need to stay.

The new compiler uses the NPM package react-tools. When installing that
you get a deprecation warning with a link to a blog article
(https://facebook.github.io/react/blog/2015/06/12/deprecating-jstransform-and-react-tools.html)
on what is going on. The blog basically says that React as a team, on Jun 12, 2015,
made the decision to work with the babel team and begin to utilize babel
as their standard build system.

Babel replaces Facebook's JSTransform, the source transformation tool
that they wrote in-house. However due to increasing demand Facebook was
just not able to keep JSTransform up to date, so they decided to utilize
another teams build system and let the babel team worry about bringing
in new JavaScript language features. Good for them, smart move in my
opinion.

It is sad though that in order to run babel you have to include five NPM
packages.

Oh well, I guess it is what it is.